Private contacts and register,login done with jwt token

+++ controllers/contact.controller.js
@@ -30,7 +30,8 @@ const createContact = asyncHandler(async(req, res) => {
 const contact=await Contact.create({
     name,
     email,
-    phone
+    phone,
+    user_id:req.user.id
 })
 
         res.status(201).json(contact)
@@ -42,22 +43,40 @@ const contact=await Contact.create({
 //Update Contact
 const updateContact = asyncHandler(async(req, res) => {
     const {id}=req.params
-    const contacts=await Contact.findByIdAndUpdate(id,req.body)
-    if(!contacts)
+    const contact=await Contact.findById(id)
+    if (!contact) {
+        res.status(404)
+        throw new Error("contact Not Found");
+        
+    }
+    if (contact.user_id.toString()!==req.user.id) {
+        res.status(403)
+        throw new Error("User don't have permission to update other contacts");
+    }
+
+    const updatedContact=await Contact.findByIdAndUpdate(id,req.body,{new:true})
+    if(!updatedContact)
     {
         res.status(404).json({message:"Contact Not found"})
     }
-    const updatedContact=await Contact.findById(id)
     res.status(200).json(updatedContact)
 })
 //Delete Contact
 const deleteContact = asyncHandler(async(req, res) => {
     const {id}=req.params
-    const contact=await Contact.findByIdAndDelete(id)
+    const contact=await Contact.findById(id)
     if (!contact) {
         res.status(404).json("Contact not found")
     }
-    res.status(200).json(contact)
+    if (contact.user_id.toString()!==req.user.id) {
+        res.status(403)
+        throw new Error("User don't have permission to delete other contacts");
+    }
+    const deletedContact=await Contact.findByIdAndDelete(id)
+    if (!deletedContact) {
+        res.status(404).json("Deleted contact not found")
+    }
+    res.status(200).json(deletedContact)
 })
 
 module.exports = {
